continuation of last commit
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main/sip/resiprocate@3162 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ SdpContents.hxx
@@ -93,8 +93,8 @@ class SdpContents : public Contents
             {
                public:
                   Origin(const Data& user,
-                         const int& sessionId,
-                         const int& version,
+                         const unsigned long long& sessionId,
+                         const unsigned long long& version,
                          AddrType addr,
                          const Data& address);
                   Origin(const Origin& rhs);
